fix(tests): remove federated worker utils tests, move the rest to `FederatedUrlParserTest`
